+++ crates/map/src/themes/animation.rs
@@ -3,7 +3,7 @@ use crate::prelude::*;
 #[derive(Serialize, Deserialize)]
 pub struct Animation {
     pub tile_type: u16,
-    pub frames_per_second: f32,
+    pub frames_per_second: f64,
     pub frames: Vec<Frame>,
 }
